Integrate @radix-ui/react-select package and refactor UI
This commit introduces the @radix-ui/react-select package into the project and updates related dependency files. It refactors the UI, especially in the Explore.tsx by replacing the className div components with the Select components from the new package. A new file, select.tsx, is added to the src/components/ui directory to make these Select components available for use across the project.